break-out Collections rendering from CatalogScreen
